Book/edit&update

+++ kamonohasi/public/js/app.js
@@ -7,13 +7,15 @@
   \************************************/
 /***/ (() => {
 
-document.getElementById('deleteUserBtn').onclick = function deleteUser() {
-  event.preventDefault();
+if (document.getElementById('deleteUserBtn') != null) {
+  document.getElementById('deleteUserBtn').onclick = function deleteUser() {
+    event.preventDefault();
 
-  if (window.confirm('本当に削除しますか？')) {
-    document.getElementById('user-delete').submit();
-  }
-};
+    if (window.confirm('本当に削除しますか？')) {
+      document.getElementById('user-delete').submit();
+    }
+  };
+}
 
 /***/ })
