+++ kamonohasi/resources/js/deleteUser.js
@@ -1,6 +1,8 @@
-document.getElementById('deleteUserBtn').onclick = function deleteUser() {
-    event.preventDefault();
-    if (window.confirm('本当に削除しますか？')){
-        document.getElementById('user-delete').submit();
+if(document.getElementById('deleteUserBtn') != null){
+    document.getElementById('deleteUserBtn').onclick = function deleteUser() {
+        event.preventDefault();
+        if (window.confirm('本当に削除しますか？')){
+            document.getElementById('user-delete').submit();
+        }
     }
 }
